refactor: extract runner story text support
