+++ Makefile
@@ -11,6 +11,7 @@ check:
 		app/services/runner_scene_characters.py \
 		app/services/runner_session.py \
 		app/services/runner_story_support.py \
+		app/services/runner_story_text.py \
 		app/services/runner_video_prompts.py
 
 story-provider:
